feat:refactor Kubernetes config
